requirements.txt
# Core streaming and data processing
kafka-python>=2.0.0
pyspark>=3.0.0

# Data handling and analysis
pandas>=2.0.0
numpy>=1.20.0
requests>=2.25.0
python-dotenv>=0.19.0

# Real-time processing
confluent-kafka>=2.0.0

# Web framework and dashboard
flask>=2.0.0,<2.3.0
flask-socketio>=5.0.0
plotly>=5.0.0
dash>=2.0.0
dash-bootstrap-components>=1.0.0

# Data visualization
matplotlib>=3.5.0
seaborn>=0.11.0
folium>=0.12.0

# Geospatial processing
geopandas>=0.10.0
shapely>=1.8.0

# Time series and anomaly detection
scikit-learn>=1.0.0
statsmodels>=0.13.0

# API and HTTP
aiohttp>=3.7.0

# Configuration and logging
pyyaml>=5.4.0
python-json-logger>=2.0.0

# Development and testing
pytest>=6.0.0
black>=22.0.0
flake8>=4.0.0 